Add lists component to one folder (Allowing to create something new)
