+++ packages/dnd/playgrounds/dflex-react-dnd/src/lists/component-based-event/readme.md
