+++ packages/dnd/playgrounds/dflex-react-dnd/src/lists/container-based-event/Core.jsx
