+++ packages/dnd/playgrounds/dflex-react-dnd/src/lists/container-based-event/readme.md
