+++ packages/dnd/playgrounds/dflex-react-dnd/src/lists/index.js
@@ -1,0 +1,3 @@
+import { ComponentBasedEvent, ContainerBasedEvent } from "./Lists";
+
+export { ComponentBasedEvent, ContainerBasedEvent };
